Acronyms; Chapter 3 to SAA Estimator Done*
Added the framework for acronyms; need to fill out at some point when I want to do formatting busy-work.

Finished section 3.1 (DEP), 3.1.1 (sDEP), and sDEP SAA estimator section.  Last one requires the actual analysis of the estimator, determining the efficacy of the estimator to determine how large \hat{\Omega} needs to be to closely approximate \Omega.  Will fill in after analysis is done, or just edit out entirely with a mention in the future work section of my conclusion.