Started 4.2
Filled out 4.2 through to the lede of the results.  The parameters table needs to be corrected and the network field figures need to be actually built and placed.

Also did some more work considering placement of figures elsewhere in the these, primarily Section 4.1.  Figures and tables go immediately after the paragraph that first references them.  If they cannot fit there, they instead go to the top of the next page.  If they there isn't space for at least 6 lines of text following the figure/table, then the figure/table gets the page to itself.  Consider these rules for the other figures (Ch1, Ch2) and double check the parameter tables of Ch4 to ensure whether they should remain alone on a page.