Section 4.3 Rework
New data caused Section 4.3 to be rewritten.  99% of the section was rewritten from scratch with the new data.  This also impacted some of the conclusions in the conclusion (Ch 5).  Await revisions/editing for in the morning, and forward to committee first thing possible.

Need to parse through figures folder again to clean useless figures; some got removed during rewrite that may not have been removed.