Beginning Final Updates
Received comments to correct my thesis in the eyes of the grad school.  All but adding a general audience abstract is complete; finish writing the general audience abstract as soon as possible.

Also considering adding an acknowledgments page.  Already added both the GAA and ack. pages, but need to fill them out.